+++ src/components/Discover/ShareCharts/index.tsx
@@ -2,30 +2,17 @@ import React, { useState, useEffect } from 'react';
 import { Box, Tabs, Tab } from '@material-ui/core';
 
 import { IProjectCardProps } from '../../../types';
+import TabPanel from './TabPanel';
 import PieChart from './PieChart';
 import Progress from '../../common/Progress';
 import service from '../../../service';
 
-interface TabPanelProps {
-  children: React.ReactNode;
-  index: any;
-  value: any;
-}
-
-function TabPanel(props: TabPanelProps) {
-  const { children, value, index } = props;
-  return (
-    <div role="tabpanel" hidden={value !== index} id={`tabpanel-${index}`}>
-      {value === index && <Box width="100%">{children}</Box>}
-    </div>
-  );
-}
-
 interface IProps {
   selectedProjects: IProjectCardProps[];
+  period: string;
 }
 function ShareCharts(props: IProps): React.ReactElement {
-  const { selectedProjects } = props;
+  const { selectedProjects, period } = props;
 
   const [currTab, setCurrTab] = useState(0);
   const [columns, setColumns] = useState<any>();
@@ -35,11 +22,11 @@ function ShareCharts(props: IProps): React.ReactElement {
       const res = await service.getSharesData({
         projectIds: selectedProjects.map((project) => project._id),
         type: 'recent',
-        period: '1w',
+        period,
       });
       setColumns(res.data);
     })();
-  }, [selectedProjects]);
+  }, [selectedProjects, period]);
 
   const handleChange = (event: any, newValue: number) => {
     setCurrTab(newValue);
